Split _updateRecipe.ts into brewableResources + deriveBatchPhases (#202)
_updateRecipe.ts had become a grab-bag: its default only derives
batch.phases (the name misled since #196) while the file also held the
shared assignment-narrowing helpers — two now-independent concerns.

Split into actions/brewableResources.ts (ResourceFor/resourcesOf/
indexedResourcesOf) and actions/deriveBatchPhases.ts (phasesFromBrewable
+ the renamed deriveBatchPhases default), delete _updateRecipe.ts, and
update the four import sites. Also refresh two stale comments.

Closes #202
