+++ packages/app/CLAUDE.md
@@ -68,10 +68,10 @@ useKbX() → IndexedDB hit? return it
 
 ## Derived batch data (`src/actions/`)
 **Purpose.** Several batch fields are **projections of the ingredients**, not edited directly — flat, tagged views rebuilt by `_updateX(batch)` actions (`Object.assign(batch, …)`).
-**Where.** `_updateShopping` → `batch.shopping`; `_updateSchedule` → `batch.schedule` (see _BatchSchedule_); `_updateRecipe(brewable, batch)` derives `batch.phases` from `brewable` at **creation**; `_projectBatchBrewable(batch)` re-derives `batch.phases` on every **edit** save (via `_updateRecipe`'s `phasesFromBrewable`). Both used to rebuild the legacy `grains/hops/yeasts/additives` arrays too, but #196 removed those arrays (they had no readers once shopping/schedule/summary moved onto the brewable), so `phases` — per-phase equipment, with no reader-side replacement yet — is all that's left to derive. `updateBatch` (edit save path), `createBatch` (instantiation).
+**Where.** `_updateShopping` → `batch.shopping`; `_updateSchedule` → `batch.schedule` (see _BatchSchedule_); `deriveBatchPhases(brewable, batch)` (`actions/deriveBatchPhases.ts`) derives `batch.phases` from `brewable` at **creation**; `_projectBatchBrewable(batch)` re-derives it on every **edit** save (both via that file's `phasesFromBrewable`). The assignment-narrowing helpers `resourcesOf`/`indexedResourcesOf` live in `actions/brewableResources.ts`. Both used to rebuild the legacy `grains/hops/yeasts/additives` arrays too, but #196 removed those arrays (they had no readers once shopping/schedule/summary moved onto the brewable), so `phases` — per-phase equipment, with no reader-side replacement yet — is all that's left to derive. `updateBatch` (edit save path), `createBatch` (instantiation).
 **How it works.** Two conventions make rebuilds safe:
 - **Reuse-by-reference** — each rebuild matches new items against the previous list by a stable key and preserves *user-owned* fields (shopping `cost`/`purchased`; schedule `completed`/`actual`). When nothing it owns changed, it returns the **previous object by reference** so the `isEqual` diff stays cheap. `_projectBatchBrewable` follows the same rule for phase equipment — `phasesFromBrewable` rebuilds each phase's kit fresh (always `completed: false`), so it re-matches the rebuilt items to the previous phase's by name and carries `completed` over rather than resetting brew-day checkoff on every save.
-- **Trigger diffing** — `updateBatch` re-runs a derivation only when a trigger field changed (`shoppingTriggers`, `scheduleTriggers`), which are now both keyed on `"brewable"` (schedule also on `"hydrometer"`) since that's where the ingredients live; editing the batch name doesn't rebuild the schedule. It still runs `_projectBatchBrewable` first to keep `batch.phases` in sync, but the triggers no longer depend on that ordering (they compare `batch.brewable` directly). `createBatch` instead builds `batch.brewable` first — a deep clone of the recipe's own brewable for a **user** recipe, `kbBrewableToBrewable(kbRecipe.brewable)` for a **kb** one (narrowed straight from the kb data's own brewable, not derived — see _Model boundary_) — then runs the pipeline once: `_updateRecipe` → `_updateShopping` → `_updateSchedule`. `batch-planning` edits `batch.brewable` directly (via `BrewableEdit` — see _Model boundary_); the other batch views now source from the brewable too — `_updateShopping`/`_updateSchedule` read their ingredients from `brewable.assignments` (via `resourcesOf`/`indexedResourcesOf`), `batch-summary` via `organicNames`, and `batch-schedule`'s write-through `path`s target `brewable.assignments[i].resource.*` (see _BatchSchedule_). The legacy ingredient arrays had no readers once that landed, so #196 removed them from **both** `Batch` and `Recipe`/`KbRecipe` (plus the kb data's top-level arrays and `KbAssignment`'s indexes into them): `_projectRecipeBrewable` is deleted (`saveRecipe` just persists the recipe), and `_projectBatchBrewable`/`_updateRecipe` are trimmed to a **phases-only** re-derive. Ingredients/equipment live only on the brewable now.
+- **Trigger diffing** — `updateBatch` re-runs a derivation only when a trigger field changed (`shoppingTriggers`, `scheduleTriggers`), which are now both keyed on `"brewable"` (schedule also on `"hydrometer"`) since that's where the ingredients live; editing the batch name doesn't rebuild the schedule. It still runs `_projectBatchBrewable` first to keep `batch.phases` in sync, but the triggers no longer depend on that ordering (they compare `batch.brewable` directly). `createBatch` instead builds `batch.brewable` first — a deep clone of the recipe's own brewable for a **user** recipe, `kbBrewableToBrewable(kbRecipe.brewable)` for a **kb** one (narrowed straight from the kb data's own brewable, not derived — see _Model boundary_) — then runs the pipeline once: `deriveBatchPhases` → `_updateShopping` → `_updateSchedule`. `batch-planning` edits `batch.brewable` directly (via `BrewableEdit` — see _Model boundary_); the other batch views now source from the brewable too — `_updateShopping`/`_updateSchedule` read their ingredients from `brewable.assignments` (via `resourcesOf`/`indexedResourcesOf`), `batch-summary` via `organicNames`, and `batch-schedule`'s write-through `path`s target `brewable.assignments[i].resource.*` (see _BatchSchedule_). The legacy ingredient arrays had no readers once that landed, so #196 removed them from **both** `Batch` and `Recipe`/`KbRecipe` (plus the kb data's top-level arrays and `KbAssignment`'s indexes into them): `_projectRecipeBrewable` is deleted (`saveRecipe` just persists the recipe), and `_projectBatchBrewable`/`deriveBatchPhases` are trimmed to a **phases-only** re-derive. Ingredients/equipment live only on the brewable now.
 **Invariants.** _None._
 **Gotchas.** ⚠️ Derived fields are **not yet migrated** (see _Batch versioning & migrations_): a batch stored before a derived field existed throws until re-derived or purged (`/?purge=true`). ⚠️ Relatedly, `schedule` is only re-derived when a `scheduleTrigger` changes, so a batch whose `schedule` was built by *older* code keeps its old write-through `path`s (`hops[i].boil`) until re-derived (any ingredient edit) or re-created — until then a brew-day boil edit writes a now-removed legacy field (`batch.hops` no longer exists as of #196) and is simply lost. Freshly created/re-derived batches carry the `brewable.assignments[...]` paths and edit correctly. The standalone Checklists screen + its `_updateChecklists` derivation were **removed**; equipment checkoff moved onto BatchSchedule phases, and the batch's second tab is now `Shopping`. (`model/recipe.ts` is now the live user-recipe model — see _Model boundary_.) `actions/_projectBatchBrewable.ts` (now phases-only) is wired into `updateBatch` instead of `state/batches.ts`'s `saveBatch` because the same edit-save path also runs the shopping/schedule trigger diff. (The recipe side had an analogous `_projectRecipeBrewable`, deleted in #196 once its output arrays were removed — `saveRecipe` now just persists the recipe.)
 **Example.** _None._
