+++ packages/app/src/model/batch.ts
@@ -120,9 +120,10 @@ export default interface Batch extends Entity {
 
     /**
      * The batch's source-of-truth brewable — cloned from the recipe (user source)
-     * or narrowed via kbBrewableToBrewable (kb source) at creation time. Projected
-     * onto the legacy fields below by createBatch/_updateRecipe; not yet read by
-     * any batch screen (see _Derived batch data_ in CLAUDE.md).
+     * or narrowed via kbBrewableToBrewable (kb source) at creation time. Shopping,
+     * schedule, and summary all derive from it, and batch-planning edits it;
+     * `deriveBatchPhases` derives `phases` from it (see _Derived batch data_ in
+     * CLAUDE.md).
      */
     brewable: Brewable;
